Still restructuring app and making sure all files are being correctly imported. Also created a test folder to test functions and database connections
