+++ backend/apis/base.py
@@ -1,0 +1,3 @@
+from web_scraper.constants import *
+
+print(f"{BASE_URL}")
