+++ backend/dependencies/security_funct.py
@@ -1,38 +1,17 @@
 from datetime import datetime, timedelta
 
-from backend.schemas.security_classes import *
-from fastapi import HTTPException, Depends, status
-from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError, jwt
-
-## https://www.fastapitutorial.com/blog/password-hashing-fastapi/
-oauth2_scheme = OAuth2PasswordBearer(
-    tokenUrl="/token"
-    )
+from schemas.security_classes import settings
+from fastapi import status
+from jose import jwt
+from db.repository.users import get_user
 
 
-fake_users_db = {
-    "johndoe": {
-        "username": "johndoe",
-        "full_name": "John Doe",
-        "email": "johndoe@example.com",
-        "hashed_password": "secret",
-        "disabled": False,
-    },
-    "poptart": {
-        "username": "poptart",
-        "full_name": "poptart Doe",
-        "email": "poptart@example.com",
-        "hashed_password": "secret3",
-        "disabled": False,
-    }    
-}
 
 # -- Token Authentication Functions
-
 def decode_token(token):
-    user = get_user(fake_users_db, token)
+    user = get_user(token)
     return user
+
 
 def create_access_token(data: dict, expires_delta: timedelta | None = None):
     to_encode = data.copy()
@@ -45,6 +24,7 @@ def create_access_token(data: dict, expires_delta: timedelta | None = None):
     
     return encoded_jwt
 
+
 def create_refresh_token(data: dict, expires_delta: timedelta | None = None):
     to_encode = data.copy()
     if expires_delta is not None:
@@ -56,50 +36,3 @@ def create_refresh_token(data: dict, expires_delta: timedelta | None = None):
     encoded_jwt = jwt.encode(to_encode, settings.JWT_REFRESH_SECRET_KEY, status.ALGORITHM)
     
     return encoded_jwt
-
-# -- Password Functions
-
-
-# -- Username Functions
-
-def authenticate_user(fake_db, username: str, password: str):
-    user = get_user(fake_db, username)
-    if not user:
-        return False
-    
-    if not Hasher.verify_password(user.hashed_password, Hasher.get_password_hash(password)):
-        return False
-    
-    return user
-
-
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
-        return UserInDB(**user_dict)
-
-
-async def get_current_user(token: str = Depends(oauth2_scheme)):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, settings.JWT_SECRET_KEY, algorithms=[status.ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = get_user(fake_users_db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-async def get_current_active_user(current_user: User = Depends(get_current_user)):
-    if current_user.disabled:
-        raise HTTPException(status_code = 400, detail = "Inactive user")
-    return current_user
