+++ backend/main.py
@@ -1,16 +1,14 @@
 from fastapi import Depends, FastAPI, Security, status
 from fastapi.middleware.cors import CORSMiddleware
-from motor.motor_asyncio import AsyncIOMotorClient
 from beanie import init_beanie
 
 
-from backend.dependencies.security_funct import get_current_active_user
-from apis.base import api_router 
-from backend.schemas.security_classes import settings
-from backend.internal import admin
-from backend.apis.version1 import security_paths, card_paths, token_paths
-from backend.web_scraper.card_page.constants import *
-from backend.schemas.card_classes import *
+from db.repository.users import get_current_active_user
+from schemas.security_classes import settings
+from internal import admin
+from apis.version1 import security_paths, card_paths, token_paths
+from web_scraper.constants import *
+from schemas.card_classes import *
 
 
 
@@ -58,25 +56,8 @@ app.add_middleware(
     allow_headers = ["*"],
 )
 
-@app.on_event("startup")
-async def app_init():
-    """
-        initialize crucial application services
-    """ 
-
-    db_client = AsyncIOMotorClient(settings.MONGO_CONNECTION_STRING).carddb
-
-    await init_beanie(
-        database=db_client,
-        document_models=[]
-    )
-
-
 @app.get("/", status_code=status.HTTP_200_OK)
 async def home():
     return {"Greeting": "Welcome to the My Hero Academia Card Game API! This is a fan-made API that any developer is free to use. The only thing I ask is that you give me some credit when you use it, and/or buy me a coffee. This carbon-based lifeform needs Java installed..."}
-
-def include_router(app):
-    app.include_router(api_router)
 
 # -- uvicorn main:app --reload
